+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed `Command#showHelp()`, it's now logs help message in console instead of returning it
+
 ## 2.0.1 (December 16, 2019)
 
 - Fixed multiline description output in help